+++ api/typescript/src/engine/kurtosis_engine_rpc_api_bindings/engine_service_grpc_web_pb.js
@@ -438,5 +438,66 @@ proto.engine_api.EngineServicePromiseClient.prototype.clean =
 };
 
 
+/**
+ * @const
+ * @type {!grpc.web.MethodDescriptor<
+ *   !proto.engine_api.GetUserServiceLogsArgs,
+ *   !proto.engine_api.GetUserServiceLogsResponse>}
+ */
+const methodDescriptor_EngineService_GetUserServiceLogs = new grpc.web.MethodDescriptor(
+  '/engine_api.EngineService/GetUserServiceLogs',
+  grpc.web.MethodType.UNARY,
+  proto.engine_api.GetUserServiceLogsArgs,
+  proto.engine_api.GetUserServiceLogsResponse,
+  /**
+   * @param {!proto.engine_api.GetUserServiceLogsArgs} request
+   * @return {!Uint8Array}
+   */
+  function(request) {
+    return request.serializeBinary();
+  },
+  proto.engine_api.GetUserServiceLogsResponse.deserializeBinary
+);
+
+
+/**
+ * @param {!proto.engine_api.GetUserServiceLogsArgs} request The
+ *     request proto
+ * @param {?Object<string, string>} metadata User defined
+ *     call metadata
+ * @param {function(?grpc.web.RpcError, ?proto.engine_api.GetUserServiceLogsResponse)}
+ *     callback The callback function(error, response)
+ * @return {!grpc.web.ClientReadableStream<!proto.engine_api.GetUserServiceLogsResponse>|undefined}
+ *     The XHR Node Readable Stream
+ */
+proto.engine_api.EngineServiceClient.prototype.getUserServiceLogs =
+    function(request, metadata, callback) {
+  return this.client_.rpcCall(this.hostname_ +
+      '/engine_api.EngineService/GetUserServiceLogs',
+      request,
+      metadata || {},
+      methodDescriptor_EngineService_GetUserServiceLogs,
+      callback);
+};
+
+
+/**
+ * @param {!proto.engine_api.GetUserServiceLogsArgs} request The
+ *     request proto
+ * @param {?Object<string, string>=} metadata User defined
+ *     call metadata
+ * @return {!Promise<!proto.engine_api.GetUserServiceLogsResponse>}
+ *     Promise that resolves to the response
+ */
+proto.engine_api.EngineServicePromiseClient.prototype.getUserServiceLogs =
+    function(request, metadata) {
+  return this.client_.unaryCall(this.hostname_ +
+      '/engine_api.EngineService/GetUserServiceLogs',
+      request,
+      metadata || {},
+      methodDescriptor_EngineService_GetUserServiceLogs);
+};
+
+
 module.exports = proto.engine_api;