+++ docs/changelog.md
@@ -15,6 +15,10 @@
 
 # 0.50.0
 ### Features
+* Created new engine's endpoint `GetUserServiceLogs` for consuming user service container logs from the logs database server
+* Added `LogsDatabaseClient` interface for defining the behaviour for consuming logs from the centralized logs database
+* Added `LokiLogsDatabaseClient` which implements `LogsDatabaseClient` for consuming logs from a Loki's server
+* Added `KurtosisBackendLogsClient` which implements `LogsDatabaseClient` for consuming user service container logs using `KurtosisBackend`
 * Created the `LogsDatabase` object in `container-engine-lib`
 * Created the `LogsCollector` object in `container-engine-lib`
 * Added `LogsDatabase` CRUD methods in `Docker` Kurtosis backend